first version of rma6 writer for MALT

+++ src/malt/data/ReadMatch.java
@@ -78,7 +78,7 @@ class ReadMatch {
         this.endRef = endRef;
     }
 
-    public double getBitScore() {
+    public float getBitScore() {
         return bitScore;
     }
